+++ dev/src/main.ts
@@ -2,6 +2,7 @@ import { createApp } from 'vue';
 import App from './App.vue';
 import router from './router';
 import { AdapterPlugin, IconsPlugin } from '@vtj/web';
+import ElementPlus from 'element-plus';
 import '@/style/index.scss';
 import './mock';
 const app = createApp(App);
@@ -12,4 +13,5 @@ app.use(AdapterPlugin, {
     // size: 'mini' /
   }
 });
+app.use(ElementPlus);
 app.mount('#app');