+++ dev/src/views/designer/renderer.vue
@@ -1,12 +1,28 @@
 <template>
   <div>
     <component :is="renderer"></component>
+    <div v-pin:[arg].modifiers="true">abc</div>
   </div>
 </template>
 <script lang="ts" setup>
-  import { reactive } from 'vue';
-  import { createRenderer, type BlockSchema } from '@vtj/pro';
+  import { reactive, ref } from 'vue';
+
+  import {
+    createRenderer,
+    type BlockSchema,
+    type NodeDirective
+  } from '@vtj/pro';
   import { XAction } from '@vtj/web';
+  const vPin = {
+    mounted(el: any, binding: any) {
+      el.innerHTML = JSON.stringify(binding);
+    },
+    updated() {
+      /* ... */
+    }
+  };
+
+  const arg = ref('top');
 
   const dsl: BlockSchema = reactive({
     name: 'Test',
@@ -25,7 +41,36 @@
         name: 'XAction',
         props: {
           label: 'ABC'
-        }
+        },
+        directives: [
+          {
+            name: 'vIf',
+            value: {
+              type: 'JSExpression',
+              value: 'true'
+            }
+          },
+          {
+            name: 'loading',
+            value: {
+              type: 'JSExpression',
+              value: 'true'
+            }
+          }
+        ] as NodeDirective[]
+      },
+      {
+        name: 'div',
+        children: 'Div Content',
+        directives: [
+          {
+            name: 'vHtml',
+            value: {
+              type: 'JSExpression',
+              value: '"<h1>aaa</h1>"'
+            }
+          }
+        ] as NodeDirective[]
       }
     ]
   });