+++ packages/coder/src/parser/template.ts
@@ -14,7 +14,7 @@ import {
   type NodeFromPlugin,
   BUILT_IN_TAGS
 } from '@vtj/core';
-import { isPlainObject, camelCase, dedupArray } from '@vtj/base';
+import { isPlainObject, camelCase, dedupArray, kebabCase } from '@vtj/base';
 import {
   isJSExpression,
   parseValue,
@@ -24,6 +24,18 @@ import {
   replaceThis,
   isJSCode
 } from '../utils';
+
+/**
+ * 内置指令
+ */
+export const BUILT_IN_DIRECTIVES = [
+  'vIf',
+  'vShow',
+  'vModel',
+  'vFor',
+  'vBind',
+  'vHtml'
+];
 
 export function parseTemplate(
   children: NodeSchema[],
@@ -35,6 +47,7 @@ export function parseTemplate(
   const nodes: string[] = [];
   let methods: Record<string, JSFunction> = {};
   let components: string[] = [];
+  const defineDirectives: string[] = [];
   let importBlocks: { id: string; name: string }[] = [];
   const slots = groupBySlot(children);
   slots.forEach((item) => {
@@ -64,9 +77,11 @@ export function parseTemplate(
         computedKeys
       );
 
-      const directives = parseDirectives(child.directives, computedKeys).join(
-        ' '
-      );
+      const directives = parseDirectives(
+        child.directives,
+        computedKeys,
+        defineDirectives
+      ).join(' ');
       const nodeChildren = child.children
         ? parseNodeChildren(
             child.children,
@@ -98,12 +113,21 @@ export function parseTemplate(
     const node = wrapSlot(item.slot, contents.join('\n'), parent?.id);
     nodes.push(node);
   });
+
   return {
     nodes,
     methods,
+    directives: directivesRegister(defineDirectives),
     components: dedupArray(components) as string[],
     importBlocks: dedupArray<{ id: string; name: string }>(importBlocks, 'id')
   };
+}
+
+function directivesRegister(directives: string[]) {
+  return dedupArray(directives).map((name) => {
+    const key = name.startsWith('v') ? name.substring(1) : name;
+    return `${key}:${name}`;
+  });
 }
 
 function groupBySlot(children: NodeSchema[] = []) {
@@ -286,10 +310,12 @@ function parsePropsAndEvents(
 
 function parseDirectives(
   directives: NodeDirective[] = [],
-  computedKeys: string[] = []
+  computedKeys: string[] = [],
+  output: string[] = []
 ) {
   const result: string[] = [];
-  const { vIf, vShow, vModels, vFor, vBind } = getDiretives(directives);
+  const { vIf, vShow, vModels, vFor, vBind, vHtml, customDirectives } =
+    getDiretives(directives);
   if (vIf) {
     result.push(`v-if="${parseValue(vIf.value, true, true, computedKeys)}"`);
   }
@@ -323,22 +349,91 @@ function parseDirectives(
       `v-for="(${item}, ${index}) in ${parseValue(vFor.value, true, true, computedKeys)}"`
     );
   }
+
+  if (vHtml) {
+    result.push(
+      `v-html="${parseValue(vHtml.value, true, true, computedKeys)}"`
+    );
+  }
+
+  if (customDirectives && customDirectives.length) {
+    customDirectives.forEach((dir) => {
+      if (!dir.name) return;
+      let leftExp = '';
+      let source = '';
+      if (isJSExpression(dir.name)) {
+        source = parseValue(dir.name, true, true, computedKeys);
+        output.push(source);
+      } else {
+        source = dir.name;
+      }
+      const name = source?.startsWith('v')
+        ? kebabCase(source)
+        : kebabCase('v-' + source);
+
+      leftExp += name;
+
+      if (dir.arg) {
+        if (isJSExpression(dir.arg)) {
+          leftExp += `:[${parseValue(dir.name, true, true, computedKeys)}]`;
+        } else {
+          leftExp += `:${dir.arg}`;
+        }
+      }
+      if (dir.modifiers) {
+        const keys = Object.keys(dir.modifiers);
+        if (keys.length) {
+          leftExp += keys.map((n) => '.' + n);
+        }
+      }
+      if (dir.value) {
+        result.push(
+          `${leftExp}="${parseValue(dir.value, true, true, computedKeys)}"`
+        );
+      } else {
+        result.push(leftExp);
+      }
+    });
+  }
   // todo: 实现others 指令
   return result;
 }
 
 function getDiretives(directives: NodeDirective[] = []) {
-  const vIf = directives.find((n) => camelCase(n.name) === 'vIf');
-  const vFor = directives.find((n) => camelCase(n.name) === 'vFor');
-  const vShow = directives.find((n) => camelCase(n.name) === 'vShow');
-  const vBind = directives.find((n) => camelCase(n.name) === 'vBind');
-  const vModels = directives.filter((n) => camelCase(n.name) === 'vModel');
+  const builtInDirectives = directives.filter((n) =>
+    BUILT_IN_DIRECTIVES.includes(n.name as string)
+  );
+
+  const customDirectives = directives.filter(
+    (n) => !BUILT_IN_DIRECTIVES.includes(n.name as string)
+  );
+
+  const vIf = builtInDirectives.find(
+    (n) => camelCase(n.name as string) === 'vIf'
+  );
+  const vFor = builtInDirectives.find(
+    (n) => camelCase(n.name as string) === 'vFor'
+  );
+  const vShow = builtInDirectives.find(
+    (n) => camelCase(n.name as string) === 'vShow'
+  );
+  const vBind = builtInDirectives.find(
+    (n) => camelCase(n.name as string) === 'vBind'
+  );
+  const vHtml = builtInDirectives.find(
+    (n) => camelCase(n.name as string) === 'vHtml'
+  );
+  const vModels = builtInDirectives.filter(
+    (n) => camelCase(n.name as string) === 'vModel'
+  );
   return {
     vIf,
     vFor,
     vShow,
     vModels,
-    vBind
+    vBind,
+    vHtml,
+    customDirectives
   };
 }