+++ packages/renderer/CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.3](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.9.2...@vtj/renderer@0.9.3) (2024-12-24)
+
+
+### Features
+
+* ✨ 数据源增加模拟数据类型 ([0a264d0](https://gitee.com/newgateway/vtj/commits/0a264d023c3a7c08ecad673e71fcd19e11a9e27d))
+
+
+
+
+
+## [0.9.2](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.9.1...@vtj/renderer@0.9.2) (2024-12-18)
+
+
+### Features
+
+* ✨ 支持自定义指令 ([8a193fd](https://gitee.com/newgateway/vtj/commits/8a193fddd211d4b69d282bdfbfe6e76c6352033f))
+
+
+
+
+
+## [0.9.1](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.9.0...@vtj/renderer@0.9.1) (2024-12-17)
+
+
+### Features
+
+* ✨ 支持定义登录页 ([c23e06f](https://gitee.com/newgateway/vtj/commits/c23e06faf7049e67525d5847925680404ff0bc76))
+* ✨ render 支持自定义插槽 ([e0ea7dd](https://gitee.com/newgateway/vtj/commits/e0ea7dd8dd0384f8e666265d9e5b9ad8168cf6ef))
+
+
+
+
+
 # [0.9.0](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.9.0-alpha.50...@vtj/renderer@0.9.0) (2024-12-16)
 
 **Note:** Version bump only for package @vtj/renderer