+++ platforms/pro/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.3](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.2...@vtj/pro@0.9.3) (2024-12-24)
+
+
+### Features
+
+* ✨ 数据源增加模拟数据类型 ([0a264d0](https://gitee.com/newgateway/vtj/commits/0a264d023c3a7c08ecad673e71fcd19e11a9e27d))
+
+
+
+
+
+## [0.9.2](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.1...@vtj/pro@0.9.2) (2024-12-18)
+
+
+### Features
+
+* ✨ v-html指令 ([189ffa7](https://gitee.com/newgateway/vtj/commits/189ffa76e8329d9a8d5c43346d1a1ff4d1377ea0))
+
+
+
+
+
+## [0.9.1](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.0...@vtj/pro@0.9.1) (2024-12-17)
+
+
+### Features
+
+* ✨ 支持定义登录页 ([c23e06f](https://gitee.com/newgateway/vtj/commits/c23e06faf7049e67525d5847925680404ff0bc76))
+
+
+
+
+
 # [0.9.0](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.9.0-alpha.50...@vtj/pro@0.9.0) (2024-12-16)